Back Trace Exception to Line That Raised It
The DBMS_UTILITY.format_error_backtrace function, added in Oracle Database 10g Release 2, is a critical subprogram to call when logging exceptions. It returns a string that traces the error back to the line on which it was raised! Note: if you re-raise an exception as it propagates up the stack, you will lose the "original" line number. The back trace function always only traces back to the most recently raised exception.